# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Amo Market Portal API",
        "description": "This is the Portal API documentation for the Amo Market project.",
        "contact": {
            "email": "amomarket.info@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/portal/shops": {
            "get": {
                "tags": [
                    "Shop"
                ],
                "summary": "Get nearby shops within a 2km radius",
                "description": "Returns a list of nearby shops based on the user's current latitude and longitude.",
                "operationId": "getNearbyShops",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "description": "Latitude of the current location",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "description": "Longitude of the current location",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Shops retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Shops Retrieved Successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 101
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Fresh Mart"
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": "shop_image.jpg"
                                                    },
                                                    "rating": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 4.5
                                                    },
                                                    "time": {
                                                        "type": "string",
                                                        "example": "10-20 mins"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Grocery delivery shop."
                                                    },
                                                    "location": {
                                                        "type": "string",
                                                        "example": "Main Street"
                                                    },
                                                    "latitude": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 12.9716
                                                    },
                                                    "longitude": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 77.5946
                                                    },
                                                    "online_status": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "status": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "shop_type_name": {
                                                        "type": "string",
                                                        "example": "Grocery"
                                                    },
                                                    "has_services": {
                                                        "type": "integer",
                                                        "example": 0
                                                    },
                                                    "distance": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 1.2
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "example": "product"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing latitude or longitude",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Current location not provided."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No shops found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No shops found within 2 km."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}